move module utilities from __init__ to utils
