+++ runway/module/serverless.py
@@ -19,8 +19,8 @@ from runway.hooks.staticsite.util import get_hash_of_files
 from .._logging import PrefixAdaptor
 from ..s3_util import does_s3_object_exist, download, ensure_bucket_exists, upload
 from ..util import YamlDumper, cached_property, merge_dicts
-from . import generate_node_command, run_module_command
 from .base import ModuleOptions, RunwayModuleNpm
+from .utils import generate_node_command, run_module_command
 
 if TYPE_CHECKING:
     from pathlib import Path
